requirements.txt
streamlit>=1.20
aiohttp>=3.8
pandas>=1.5
numpy>=1.21
python-dateutil>=2.8
fpdf2>=2.5
altair>=4.2

